backend/requirements.txt
# base.txt
# Core Django & REST framework
Django==5.1.4
djangorestframework==3.15.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
django-filter==24.3 # 支持项目复杂的查询
drf-nested-routers==0.94.1
drf-polymorphic==1.0.0  # 支持多态序列化器
django-polymorphic==3.1.0 # 支持多态模型

# Database
psycopg[binary,pool]==3.2.4  #支持异步
psycopg2-binary==2.9.10 #需要保留用来支持celery
SQLAlchemy==2.0.36



# Storage
django-storages==1.14.4
cos-python-sdk-v5==1.9.33
aiobotocore==2.21.1
boto3==1.37.1
aioboto3==14.1.0 # 异步boto3


# API Documentation
drf-spectacular==0.28.0
PyYAML==6.0.2

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.31.0
urllib3==2.1.0
bcrypt==4.2.0

# Async Support
aiohttp==3.10.10
anyio==4.7.0
asyncio==3.4.3

# Time and Date handling
pytz==2024.2
tzdata==2025.2

# LLM Core
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-community == 0.3.13


redis==5.2.1
celery==5.3.6
django-redis==5.4.0
django_celery_beat==2.7.0
django_celery_results==2.5.1
# Celery 
kombu==5.5.3  # 支持使用数据库作为消息代理

# 文档处理
mammoth==1.9.0
python-docx==1.1.2

python-magic==0.4.27



# Channels for WebSocket support
channels==4.2.2
daphne==4.1.2
channels-redis==4.2.1
websockets==15.0.1

# pypandoc 3.6.4

# 腾讯云短信
tencentcloud-sdk-python-common==3.0.1375
tencentcloud-sdk-python-sms==3.0.1375
tencentcloud-sdk-python==3.0.1375

# -----------   development.txt
#-r base.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.3

# Debugging & Development Tools
ipython==8.12.3
ipykernel==6.29.0
debugpy==1.8.0
jupyter==1.1.1

jupyter-client==8.6.0
jupyter-core==5.7.1
jupyterlab==4.3.4

# Code Quality & Analysis
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development Utilities
watchfiles==0.24.0
pip-tools==7.3.0
python-json-logger==3.2.1  # Useful for development debugging and log analysis

# 文件类型检测 for windows

gunicorn==21.2.0
